+++ src/App.tsx
@@ -4,12 +4,15 @@ import './styles/theme-colors.css';
 import { BrowserRouter } from 'react-router-dom';
 import { AppRoutes } from './routes/appRoutes';
 import { MessageContainer } from './components/messageContainer';
+import { AuthProvider } from './features/auth/context/authProvider';
 
 export function Index() {
   return (
-    <BrowserRouter>
-      <MessageContainer />
-      <AppRoutes />
-    </BrowserRouter>
+    <AuthProvider>
+      <BrowserRouter>
+        <MessageContainer />
+        <AppRoutes />
+      </BrowserRouter>
+    </AuthProvider>
   );
 }
