+++ src/features/auth/services/authRegister.ts
@@ -1,4 +1,4 @@
-import { api } from './axios';
+import { api } from '../../../services/axios';
 
 export async function authRegister(
   name: string,
