+++ src/pages/register/index.tsx
@@ -1,27 +1,11 @@
 import { Button } from '../../components/button';
 import { Input } from '../../components/input';
 import { FormTemplate } from '../../template/formTemplate';
-import { authRegister } from '../../services/authRegister.ts';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { useForm } from 'react-hook-form';
-import { schema } from '../../validations/register/schema';
-import type { FormData } from '../../validations/register/schema.ts';
+import { useRegister } from '../../features/auth/hooks/useRegister.ts';
 
 export function Register() {
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors, isSubmitting },
-  } = useForm<FormData>({
-    resolver: yupResolver(schema),
-  });
-
-  async function onSubmit(data: FormData) {
-    const success = authRegister(data.name, data.email, data.password);
-
-    if (await success) return reset();
-  }
+  const { register, handleSubmit, onSubmit, errors, isSubmitting } =
+    useRegister();
 
   return (
     <FormTemplate onSubmit={handleSubmit(onSubmit)}>
